backend/requirements.txt
Django==4.2
djangorestframework==3.15.0
djangorestframework-simplejwt[crypto]==5.3.1
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
agora-token-builder==1.5.0
scikit-learn==1.3.2
